move popup into link class
